Move pipeline writers into the existing "plugins" folder
